==> 05-results.Rmd <==
# Results
```{r}
library(ggplot2)
library(dplyr)
library(latex2exp)
library(tidyverse)
library(zoo)
library(ncdf4)
library(animation)
library(viridis)
library(scales)
library(ggthemes)
library(raster)
library(reshape)
library(reshape2)
```
## Main pattern of climate change 


The first part aims to illustrate the main pattern of the Earth's surface temperature for the past two to three centuries. The global warming is definitely an important topic when it comes to the area of climate change. We first take a look at the global land-ocean temperature index since 1880. The temperature index represents the change in global surface temperature relative to 1951-1980 average temperatures. Despite some ups and downs, we can see that Earth’s surface continues to significantly warm, with recent global temperatures being the hottest in the past 2,000-plus years.

```{r fig.height = 7, fig.width = 10, fig.align = 'center'}
yearly_temp <- read.table('data/clean/graph.txt', header = TRUE)

yearly_temp %>% ggplot(aes(x = Year, y=No_Smoothing)) + geom_smooth(method = 'loess', span = 0.2, formula = y~x, se = TRUE) +
  geom_point(color = 'black', size = 0.5, alpha = 0.8)  + theme_grey(16) +
  xlab("Year") + ylab("Temperature Index") +
  theme(plot.title = element_text(size = 15, face = "bold"))

```


The global warming has indeed impacted significantly on each and every continent for the past century. The following graphs show the pattern of the average temperature since 1950.

```{r fig.height = 7, fig.width = 10, fig.align = 'center'}
temp_by_cont <- read.table('data/clean/temp_by_cont.txt', header =TRUE, sep='\t', row.names = 1)
colnames(temp_by_cont) = c(1950, 1960, 1970, 1980, 1990, 2000, 2010, 2020)
ggplot(melt(t(temp_by_cont)), aes(Var1, value)) +  
  geom_label(aes(label = Var1)) +  
  geom_line(color = "blue") + facet_wrap(~ Var2, scales="free") + xlab('Year') + 
  ylab('Average Temperature')
```
To take a closer look on what has happened around the world in terms of temperature change, the following plots show the temperature in Celsius globally in a world map currently compared to 60 years ago. We can see that many places have become hotter based on the redness of each region.

```{r}
first <- raster( "data/clean/1960_tmax.gri")
third <- raster( "data/clean/2020_tmax.gri")
```

```{r fig.height = 6, fig.width = 11, fig.align = 'center' }
plot(first, col=rev(heat.colors(48)), main = 'Avg Temperature at 1960 Spring', ylim = c(-100, 100), zlim = c(-50,50), xlab = 'Latitute', ylab = 'Longtitude')

plot(third, col=rev(heat.colors(48)), main = 'Avg Temperature at 2020 Spring', ylim = c(-100, 100), zlim = c(-50,50), xlab = 'Latitute', ylab = 'Longtitude')
```

Finally, to wrap up what we found globally, we plot the following heat map by taking a closer look at the temperature change pattern monthly. The plot captures the average temperature difference with respect to the global mean value since 1950. 

```{r fig.height=10, fig.width=12, fig.align = 'center'}
temp_dat_monthly = read.csv('data/clean/temp_dat_monthly.csv', header = TRUE)
gg <- ggplot(temp_dat_monthly, aes(x=year, y=month, fill=anomaly))
gg <- gg + geom_tile(color="white", size=0.1)
gg <- gg + scale_fill_viridis(name="Difference from \nAverage in °C",option="inferno")
#plot.title = 'Average World Temperature since 1850'
gg <- gg + ggtitle('World Temperature Difference from Avgerage since 1850') 
gg <- gg + labs(x=NULL, y=NULL)

gg <- gg + theme_tufte(base_family="Helvetica")

dgr_fmt <- function(x, ...) {
  parse(text = paste(x, "", sep = ""))
}

a <- dgr_fmt(seq(1850,2015, by=30))
gg <- gg +
  coord_cartesian(xlim = c(1850,2015)) +
  scale_x_continuous(breaks = seq(1850,2015, by=30), labels = a) +
  scale_y_continuous(expand = c(0, 0), 
                     breaks = c(1,2,3,4,5,6,7,8,9,10,11,12),
                     labels = c("Jan", "Feb", "Mar", "Apr",
                                "May", "Jun", "Jul", "Aug", "Sep",
                                "Oct", "Nov", "Dec"))
gg <- gg + ggtitle('Average World Temperature since 1850') 
gg <- gg + theme(plot.title = element_text(hjust = 0.5))
gg <- gg + theme(axis.ticks=element_blank())
gg <- gg + theme(axis.text=element_text(size=7)) 
gg
```




## Main causes of climate change
In the past more than a hundred years, the global land-ocean temperature kept increasing. So we plan to figure out what leads to the current warming trend. In the following discussion, we explore the effects of natural factors and human activities on temperature.

### Natural causes
Changes of the Earth's orbit around the sun and the tilt of the Earth's axis define the sequence of glaciation periods and warm periods, and influence the climate on Earth by changing the energy the Earth could receive from the sun. Thus the metrics of Earth's orbit and the solar irradiation are taken into account for the natural causes of climate change. In 1930s, Milutin Milankovitch introduced the concept of Milankovich cycles to describe the changes of Earth's orbit, which includes eccentricity to describe the shape of Earth's orbit, obliquity and precession to describe the Earth's axis. Eccentricity = (aphelion-perihelion)/(aphelion+perihelion). Obliquity is the angle Earth's axis of rotation is tilted as it travels around the Sun. And insolation, which denotes the daily mean solar irradiation at the top of Earth's atmosphere, is introduced to show the irradiation the Earth receives from the sun. The following shows the changes of eccentricity, obliquity and insolation in the past around five thousand years from 2800s BC to now. 




```{r fig.height = 7, fig.width = 10}
orbit <- read.table("data/raw/milankovitch.data.txt", header = TRUE, skip = 7, sep = "\t")
orbit_5k <- orbit %>% dplyr::select(time, eccentricity, obliquity, insolation) %>% transform(time = time + 2021) %>% filter(time <= 2021)
orbit_pivot <- orbit_5k %>% pivot_longer(cols = !time, names_to = "metric", values_to = "value")
orbit_pivot$metric <- orbit_pivot$metric %>% fct_relevel("eccentricity","obliquity","insolation")
labels <- c(`eccentricity`="Eccentricity", `obliquity`="Obliquity (radians)", `insolation` = "Insolation (W/m^2)")
orbit_pivot %>% ggplot(aes(x = time, y=value)) + geom_smooth(formula = y~x, se = TRUE) +
  geom_point(color = 'black', size = 0.5, alpha = 0.5) + 
  facet_wrap(~metric, scales = "free", ncol = 1, labeller = as_labeller(labels)) + theme_grey(16) +
  ggtitle("Eccentricity, Obliquity, Insolation") + xlab("Year") + ylab("") +
  theme(plot.title = element_text(size = 15, face = "bold"))
```

In the past five thousand year period, all three metrics fluctuates periodically and there is no obviously abnormal curve appears in the recent 200 years, when the global warming occurred, compared to the curve before. It seems that the value of eccentricity increases over a long time span, it is more reasonable to regard it as a normal fluctuation in a long period rather than a sudden change in short term, considering the eccentricity of Earth has a period of about a hundred thousand year.

Then we focus more on the recent 200 years. 

```{r fig.height = 7, fig.width = 10}
orbit2 <- orbit %>% dplyr::select(time, eccentricity, obliquity, insolation) %>% transform(time = time + 2021) %>% filter(time <= 2021 & time >= 1800)
orbit_pivot2 <- orbit2 %>% pivot_longer(cols = !time, names_to = "metric", values_to = "value")
orbit_pivot2$metric <- orbit_pivot2$metric %>% fct_relevel("eccentricity","obliquity","insolation")
orbit_pivot2 %>% ggplot(aes(x = time, y=value)) + geom_smooth(formula = y~x, se = TRUE) +
  geom_point(color = 'black', size = 0.5, alpha = 0.8) + 
  facet_wrap(~metric, scales = "free", ncol = 1, labeller = as_labeller(labels)) + theme_grey(16) +
  ggtitle("Eccentricity, Obliquity, Insolation") + xlab("year") + ylab("") +
  theme(plot.title = element_text(size = 15, face = "bold"))
```

In the short term, both obliquity and isolation have relatively smooth cycle changes while eccentricity has a declining trend. 

To have a better insight of the relationship between these natural causes and temperature, we use the data from 1880 to draw the scatter plots and make linear fitting. In the plot, the color of points denotes the corresponding year.
```{r fig.height = 7, fig.width = 10}

temp_yearly <- read.table("data/clean/temp_yearly.csv", header = TRUE, sep = ",")
temp_orbit_pivot = temp_yearly %>% inner_join(orbit_pivot, by = c("Year" = "time")) %>% drop_na
temp_orbit_pivot %>% ggplot(aes(value, temp, color = Year)) + 
  geom_point(alpha = 1, stroke = 0, size = 2) + 
  geom_smooth(method = "lm", se=FALSE, color="red") + 
  scale_fill_gradient() +
  facet_wrap(~metric, scales = "free", ncol = 2, labeller = as_labeller(labels)) +
  ggtitle("Natural Causes vs Global Temperature (1880 - 2020)") + xlab("") + ylab("Temperature (C)") +
  theme(plot.title = element_text(size = 15, face = "bold"))
```

From the plot between eccentricity and temperature, we could observe that there is a negative correlations between eccentricity and temperature. The points that gather at the bottom right corner are dark blue, which means in the early time, the temperature is lower while the value of eccentricity is larger, coinciding with the time series plots above. And for points in light blue, they are more likely to gather at top and middle of the left side. It shows that during the nearest time, the value of eccentricity is smaller than the previous but has a relatively stable value, while the temperature still keeps going up. So it seems that the change of Earth's orbit has influenced the temperature to some degree, but it is not the main reason for the continuous warmer weather, at least in the recent 50 years when the value of eccentricity does not change a lot. And we also need to notice that in the other two scatter plots that show relationships of temperature with obliquity and insolation, there is no obvious linear correlations between two variables, indicating that changes of the tilt of Earth's axis and the amount of solar irradiation does not play an important role in warming trend. As mentioned above, the change of Earth's orbit influences climate change by changing the solar irradiation. Although the eccentricity has an obvious change, the irradiation received by Earth keeps fluctuating between a stable value in the recent 200 years, which could be explained by the claim that there are other factors that offset the effect of Earth's orbit. Based on the analysis above, we could hardly draw to the conclusion that there are strong causal relationship between natural factors and global warming.

### Human causes
The accumulation of greenhouse gas in the atmosphere will make the heat trapped on the surface of Earth, thus lead to the temperature increase on Earth. Since the Industrial Revolution, deforestation, fossil fuel use and other human activities have led to the huge increase in green house gas. Although there are some natural activities that can also influence the gas in atmosphere, the abnormal increase should be mainly attributed to human. So we collect the global mean mole fraction of four frequently analyzed green house gas, including CH$_4$, CO$_2$, N$_2$O, SF$_6$, to explore their relationships with current climate change.

In the following plot, we show how the gas concentration of four green house gases varied monthly with time in the past years. The start time of each plot is based on the time National Oceanic and Atmospheric Administration (NOAA) started to record the global data of the gas.

```{r fig.height = 7, fig.width = 10}
gh_gas <- read.table("data/clean/gh_gas_monthly.csv", header = TRUE, sep = ",") %>% transform(time = as.yearmon(time))
gh_gas_pivot <- gh_gas %>% pivot_longer(cols = !c(time, decimal), names_to = "gas", values_to = "value") %>% drop_na
labels_gas <- c(`CH4`="CH4 (ppb)", `CO2`="CO2(ppm)", `N2O`="N2O (ppb)",`SF6`="SF6 (ppt)")
gh_gas_pivot %>% ggplot(aes(x = time, y=value)) + geom_smooth(method = 'loess', formula = y~x, se = TRUE) +
  geom_point(color = 'black', size = 0.5, alpha = 0.8) + facet_wrap(~gas, scales = "free", labeller = as_labeller(labels_gas)) + theme_grey(16) +
  ggtitle("Mole Fraction of Greenhouse Gas") + xlab("Month") + ylab("Gas Concentration") +
  theme(plot.title = element_text(size = 15, face = "bold"))
```

In addition to the periodical fluctuations, the trends for mole fraction of all four gases are increasing. Except that the curve for CH$_4$ has a platform from 2000 to 2007, other three curves show slightly faster growth than linear growth. For Further analysis, the plot below shows the growth rate of four gases. To eliminate the influence of periodical fluctuations on the understanding of the whole pattern of trend, we use the yearly data to display the changes.
```{r fig.height = 7, fig.width = 10}
gas_rate <- read.table("data/clean/gas_rate_yearly.csv", header = TRUE, sep = ",")
gas_rate_pivot <- gas_rate %>% pivot_longer(cols = !year, names_to = "gas", values_to = "rate") %>% drop_na
gas_rate_pivot %>% ggplot(aes(x = year, y=rate)) + 
  geom_col(alpha = 0.8) + expand_limits(y = 0) +
  facet_wrap(~gas, scales = "free") + theme_grey(16) +
  ggtitle("Growth Rate of Greenhouse Gas Concentration") + xlab("Year") + ylab("Growth Rate") +
  theme(plot.title = element_text(size = 15, face = "bold"))
```

Although from 1999 to 2006, the mole fraction of CH$_4$ displays a negative growth or a small growth rate, in other years from 1880s to now, its growth rate is much larger than other gases. The growth rates of other three gases are relatively small, but show a continuous increasing trend. Then we analyze the relationships between the gas concentration in the atmosphere and global temperature by drawing scatter plots and hexagonal heatmaps. Every point in scatter plots denotes mean temperature and mean mole fraction in one month from Jan 2001 to Aug 2021.The color indicates to the time in scatter plots while indicates to the density of data points in hexagonal heatmaps. 
```{r fig.height = 7, fig.width = 10}
temp_monthly <- read.table("data/clean/temp_monthly.csv", header = TRUE, sep = ",") %>% transform(time = as.yearmon(time))
temp_gas_pivot = temp_monthly %>% inner_join(gh_gas_pivot, by = c("time" = "time")) %>% filter(time >= as.yearmon("Jan 2001")) %>% dplyr::rename(Time = decimal)
temp_gas_pivot %>% ggplot(aes(value, temp, color = Time)) + 
  geom_point(alpha = 1, stroke = 0, size = 2) + 
  geom_smooth(method = "lm", se=FALSE, color="red") + 
  scale_fill_gradient() +
  facet_wrap(~gas, scales = "free", labeller = as_labeller(labels_gas)) +
  ggtitle("Gas Concentration vs Global Temperature") + xlab("Gas Concentration") + ylab("Temperature (C)") +
  theme(plot.title = element_text(size = 15, face = "bold"))
```
```{r fig.height = 7, fig.width = 10}
temp_gas_pivot %>% ggplot(aes(value, temp)) + 
  geom_hex(bins = 20) + 
  scale_fill_gradient(low = "#56B1F7", high = "#132B43") +
  facet_wrap(~gas, scales = "free" , labeller = as_labeller(labels_gas)) + theme_grey(16) +
  ggtitle("Gas Concentration vs Global Temperature") + xlab("Gas Concentration") + ylab("Temperature (C)") +
  theme(plot.title = element_text(size = 15, face = "bold"))
```

In the scatter plots, we can observe obvious positive correlations between mole fraction and temperature for four gases. And in all four plots, the points with darker colors, which means earlier times, are more likely to gather at the left bottom corner, while the lighter blue points tend to gather at the right top corner. The scatter plots displays a pattern coinciding with the analysis above that the emissions of greenhouse gas are increasing as time goes by, and at the same time, the global temperature keeps an increasing trend. Then from the hexagonal heatmaps, we observe that the points at the left bottom corner have a higher density, while points in the center and at the right top corner of plots are relatively sparse. Considering that the time is uniformly distributed, it indicates that the growth rates for both gas and temperature are higher in more recent years. From the analysis above, we find that the greenhouse gas concentration and temperature have similar behaviors in recent years. So it is of high possibility that the accumulation of greenhouse gas in the atmosphere is one of the main reason for the global warming.

## Main effects of climate change

In this part, we want to show the impacts of global warming to warn people to pay attention to climate changes. Rising temperatures have led to the gradual melting of a large number of glaciers, causing sea levels to rise and threatening coastal areas. The knock-on effects are: larger water areas --> more evaporation --> longer rainy seasons --> more flooding. Some scholars believe that global warming will lead to more extreme weather such as droughts and typhoons.

### ice sheets losing mass

The Antarctic and Arctic are key and sensitive areas for global climate change. And the Greenland and Antarctic ice sheets are the largest ice bodies in the world. So the evolution of these two ice sheets plays an important role in the global climate system. When the north and south poles are warming at a rate much higher than the global average, what pattern can we discover?

The melting Antarctic ice sheet and Greenland ice sheet:

The Antarctic ice sheet is formed by the pressing of snow on the Antarctic continent over millions of years. It covers an area of about 14 million square kilometers and has an average thickness of 2,160 meters. It accounts for about 70% of the world's total freshwater reserves.
Greenland is covered with dense snow and ice over the years, and is often referred to as the "last ice zone" of the Arctic Ocean. In the summer of 2020, sea ice has become thinner due to long-term climate change, and strong summer winds have blown away the sea ice. A vast sea surface has appeared here. Scientists have found that under the background of global warming, thinning sea ice is more vulnerable under abnormal climatic conditions. For all kinds of life in ecologically fragile areas, climate change may really mean the demise of their homes.

```{r}
library(gridExtra)
df <- read.table("data/clean/ice_sheet_antarctica_mass_200204_202109.txt",
                 skip=31,
                 col.names = c('TIME', 'antarctica', 'antarctica_sigma'))

df$antarctica_sigma <- df$antarctica_sigma+df$antarctica
df$mass_antarctic <-df$antarctica

p1 <- df %>% ggplot() + 
    geom_point(aes(x=TIME, mass_antarctic), color = 2) + 
    stat_smooth(aes(x=TIME, mass_antarctic), method = 'loess')+
    ggtitle("mass of the ice sheet over time (Antarctic)")

dt <- read.table("data/clean/ice_sheet_greenland_mass_200204_202109.txt",
                 skip=31,
                 col.names = c('TIME', 'greenland', 'greenland_sigma'))

dt$greenland_sigma <- dt$greenland_sigma+dt$greenland
dt$mass_greenland <-dt$greenland

p2 <- dt %>% ggplot() + 
    geom_point(aes(x=TIME, mass_greenland), color = 2) + 
    stat_smooth(aes(x=TIME, mass_greenland), method = 'loess')+ 
    ggtitle("mass of the ice sheet over time (Greenland)")

grid.arrange(p1,p2,ncol=1)
```
From the two graphs above, we can clearly see that the mass of the ice sheets in Antarctic and Greenland is negatively correlated with time. Since 2002, the quality of these two ice sheets has been losing more and more rapidly.
From 2002 to 2021, the cumulative ice loss in Antarctica was 2693.60 Gigatonnes; the corresponding number on Greenland Island was 5144.58 Gigatonnes.

The graphss shows that since the 21st century, the Greenland and Antarctic ice sheets have lost a large amount of ice at a rapid growth rate. Compared with the first 10 years of 2002-2011, the large-scale loss of the Greenland ice sheet in 2012-2021 has doubled, and the Antarctic ice sheet has more than tripled.

The msss of the Greenland Ice Sheet increased in 2002, but since 2003, the mass start to decrease and the rate of mass losing has been increasing. This is the same for the Antarctic ice sheet, it was still growing in 2002, and since then it has been losing mass every year. This is mainly due to the thinning of the glaciers at the main outlet of the western Antarctic ice sheet and the loss of mass of the Antarctic Peninsula ice sheet. The floating ice sheet has also become thinner.

### global sea level rising

In the previous section, we discussed the melting of ice sheets due to rising temperatures in recent years. This will further increase the sea level, an impact that we cannot ignore. Of course, there are many factors that cause the sea level to rise, not only because of the melting of ice sheets, but also because the upper seawater heats up and the volume of the ocean is expands, and so on.

Why do we explore the sea level pattern? i.e Why sea level matters?

Sea level rise is a slow-onset natural disaster to human survival and economic development. Since it is slow onset, it is often ignored by people, thinking that a few millimeters of increase per year is not a danger. In fact, this kind of disaster is cumulative and gradual. Coupled with the impact of land subsidence in many coastal areas, sea level rise may reach 1 meter or more in a century. This is undoubtedly a disaster for people in many places. And just now, it has brought harm to residents in coastal areas. It makes the occurrence of catastrophic storm surges in coastal areas more frequent, floods intensify, coasts are receded, coastal lowlands are eroded, water in coastal areas is polluted, farmland is salinized, tidal range increases, wave action is strengthened, and coastal areas are weakened. The requirement of building stronger dams will force designers to improve engineering design standards and increase project funding. It will also aggravate the intrusion of seawater in the estuary, increase the difficulty of sewage discharge, and destroy the ecological balance.

```{r}
tt <- read.table("data/clean/sea_level_GMSL_TPJAOS_5.1_199209_202108.txt", skip = 48)
head(tt)

tt %>% ggplot(mapping = aes(x = V3, y = V6, group = 1)) + 
  geom_line(color=4) + xlab('Year') +
  labs(x = "Time", y = "Global sea level ", title = "Global sea level change over time") 

```

Using data from NASA Goddard Space Flight Center, we drew the graph above. We can see that the sea level is roughly positively correlated with time. From 1993 to 2021, the sea level has risen by 95.87mm.
The rate of sea level rise is accelerating: From 1993 to 2010, the rate of sea level rise was about 2.5mm per year. However, from 2011 to 2021, the rate of sea level rise was about 4.3mm per year. This almost doubled.
Moreover, it is worth noting that from the graph, we can see that the sea level actually rises with fluctuations. In the short term, sometimes it has a temporary downward trend. This seems to contradict our argument. However, in fact, we have explained in the previous part that there are many reasons for sea level changes, sometimes not only because of climate changes, but also may be affected by human actions. For example, one of the possible factors that caused the sea level to fall in the graph is global dam projects.

### Severe storms and extreme events

As we already known from the previous part, climate changes lead to ice sheets mass loss, then induce the rising of global sea level, which contribute to flooding in coastal areas. However, it seems that this is only one of the extreme events that could be triggered by climate change. 
This year, the United Nations issued a survey report saying that today’s extreme weather outbreaks are related to the earth’s climate change, especially global warming. Anyone who has studied geography knows that a relatively stable climate zone and wind belt has formed on the earth due to the influence of factors such as latitude, the deflection of the earth, and uneven heating of the earth’s surface. The density of high-pressure air masses tends to flow toward low-pressure air masses, while the specific heat of ocean water is lower than that of land rocks. It is easier to form air masses with higher pressure over the ocean. Coupled with the influence of the deflection force of the earth, the earth has formed: polar east wind and prevailing west wind (Westerly zone), northeast monsoon, and southeast monsoon. These wind belts happen to be able to carry a large amount of water vapor from the ocean to the land. The west wind belt has more precipitation throughout the year, and the northeast, southeast monsoon belt is mainly rainy and hot in the same period. However, global warming has risen by more than 1°C compared with the last century, with the melting of extremely low glaciers and rising sea levels. However, this impact is still relatively limited. The biggest impact is the further warming of the earth, and the uneven heating between the earth's land, ocean and atmosphere will become more serious, which may cause the original climate zone and wind belt to shift. Previously, the extreme high temperature in North America was the effect of the subtropical high pressure. In the high-pressure air mass, the atmospheric subsidence caused the lack of precipitation, the lack of internal flow, and the difficulty of the external air mass to intrude. The continuous sunny days caused the ground to heat the atmosphere, which in turn made the atmosphere hotter. Both El Niño and La Niña are caused by uneven heating of the earth and abnormal temperature in local areas.

```{r}
library(readr)
library(data.table)
library(tidyverse)

dir = "data/clean/stormevents_2000_2021"
file_list = list.files(path = dir, pattern = "*.csv$",recursive = TRUE,full.names = TRUE)  
store_csv = "data/clean/stormevents_2000_2021_new.csv" #paste(dir,"new.csv")      

for(i in 1:length(file_list))     
{
    df = fread(file = file_list[i],encoding = 'UTF-8')        

    write_csv(df,path = store_csv,append = TRUE, col_names = FALSE)              
    
}

stormevents_2000_2021_new <- read_csv("data/clean/stormevents_2000_2021_new.csv")
colnames(stormevents_2000_2021_new)

tmp <- stormevents_2000_2021_new %>% group_by(`2000`) %>% count() 
tmp %>% ggplot() + geom_bar(aes(x=as.factor(`2000`), y= n, fill=as.factor(`2000`)), stat = 'identity') + 
  coord_flip() + labs(x="Time")

```
  
Observers from the National Weather Service kept a comprehensive record describing 48 different types of serious events, from localized thunderstorms, tornadoes and flash floods to regional events such as hurricanes, snowstorms and winter storms. We have counted the number of natural disasters in the United States each year from 2000 to 2021. What puzzles us is that statistics have found that the number of natural disasters does not have a clear upward trend, but rather fluctuates. Among them, the number in 2011 was the largest, almost 1.5 times that of other years. Although the number of disasters in 2021 is the least, it is not of high reference value, because 2021 has not yet ended and the data for this year is incomplete. Generally speaking, although the number of natural disasters is slowly increasing, they are very unstable and there is no obvious pattern. This leaves us room to think, does climate change really have a great impact on extreme weather?
